+++ core/store/orm/config.go
@@ -837,6 +837,11 @@ func (c Config) SessionTimeout() models.Duration {
 	return models.MustMakeDuration(c.getWithFallback("SessionTimeout", parseDuration).(time.Duration))
 }
 
+// StatsPusherLogging toggles very verbose logging of raw messages for the StatsPusher (also telemetry)
+func (c Config) StatsPusherLogging() bool {
+	return c.getWithFallback("StatsPusherLogging", parseBool).(bool)
+}
+
 // TLSCertPath represents the file system location of the TLS certificate
 // Chainlink should use for HTTPS.
 func (c Config) TLSCertPath() string {
@@ -1061,4 +1066,8 @@ func (ll LogLevel) ForGin() string {
 	default:
 		return gin.ReleaseMode
 	}
+}
+
+func parseBool(s string) (interface{}, error) {
+	return strconv.ParseBool(s)
 }